Add AddressSettings and SkillsSettings components to Profile for improved user settings management
